Draft of redesign of octree and peakfinder applications calling separate driver.
