+++ geoapps/octree_creation/__init__.py
@@ -5,5 +5,6 @@
 #  geoapps is distributed under the terms and conditions of the MIT License
 #  (see LICENSE file at the root of this source code package).
 
-from .application import OctreeMesh
+from .constants import app_initializer
+from .driver import OctreeDriver
 from .params import OctreeParams
