+++ geoapps/octree_creation/constants.py
@@ -5,10 +5,10 @@
 #  geoapps is distributed under the terms and conditions of the MIT License
 #  (see LICENSE file at the root of this source code package).
 
+from copy import deepcopy
 from uuid import UUID
 
-from geoh5py.objects import Curve
-from geoh5py.workspace import Workspace
+from geoh5py.ui_json.constants import default_ui_json as base_ui_json
 
 defaults = {
     "title": "octree Mesh Creator",
@@ -21,88 +21,82 @@ defaults = {
     "vertical_padding": 1000.0,
     "depth_core": 500.0,
     "ga_group_name": "Octree_Mesh",
-    "run_command": "geoapps.octree_creation.application",
+    "run_command": "geoapps.octree_creation.driver",
     "run_command_boolean": False,
     "monitoring_directory": None,
     "workspace_geoh5": None,
     "conda_environment": "geoapps",
     "conda_environment_boolean": False,
 }
-validations = {}
-default_ui_json = {
-    "title": "octree Mesh Creator",
-    "geoh5": None,
-    "objects": {
-        "enabled": True,
-        "group": "1- Core",
-        "label": "Core hull extent",
-        "main": True,
-        "meshType": [
-            "{202C5DB1-A56D-4004-9CAD-BAAFD8899406}",
-            "{6A057FDC-B355-11E3-95BE-FD84A7FFCB88}",
-            "{F26FEBA3-ADED-494B-B9E9-B2BBCBE298E1}",
-        ],
-        "value": None,
-    },
-    "u_cell_size": {
-        "enabled": True,
-        "group": "2- Core cell size",
-        "label": "Easting (m)",
-        "main": True,
-        "value": 25.0,
-    },
-    "v_cell_size": {
-        "enabled": True,
-        "group": "2- Core cell size",
-        "label": "Northing (m)",
-        "main": True,
-        "value": 25.0,
-    },
-    "w_cell_size": {
-        "enabled": True,
-        "group": "2- Core cell size",
-        "label": "Vertical (m)",
-        "main": True,
-        "value": 25.0,
-    },
-    "horizontal_padding": {
-        "enabled": True,
-        "group": "3- Padding distance",
-        "label": "Horizontal (m)",
-        "main": True,
-        "value": 1000.0,
-    },
-    "vertical_padding": {
-        "enabled": True,
-        "group": "3- Padding distance",
-        "label": "Vertical (m)",
-        "main": True,
-        "value": 1000.0,
-    },
-    "depth_core": {
-        "enabled": True,
-        "group": "1- Core",
-        "label": "Minimum Depth (m)",
-        "main": True,
-        "value": 500.0,
-    },
-    "ga_group_name": {
-        "enabled": True,
-        "group": None,
-        "label": "Name:",
-        "value": "Octree_Mesh",
-    },
-    "run_command": "geoapps.octree_creation.application",
-    "run_command_boolean": {
-        "value": False,
-        "label": "Run python module ",
-        "tooltip": "Warning: launches process to run python model on save",
-        "main": True,
-    },
-    "monitoring_directory": None,
-    "conda_environment": "geoapps",
-    "conda_environment_boolean": False,
-}
+
+default_ui_json = deepcopy(base_ui_json)
+default_ui_json.update(
+    {
+        "title": "octree Mesh Creator",
+        "geoh5": None,
+        "objects": {
+            "enabled": True,
+            "group": "1- Core",
+            "label": "Core hull extent",
+            "main": True,
+            "meshType": [
+                "{202C5DB1-A56D-4004-9CAD-BAAFD8899406}",
+                "{6A057FDC-B355-11E3-95BE-FD84A7FFCB88}",
+                "{F26FEBA3-ADED-494B-B9E9-B2BBCBE298E1}",
+            ],
+            "value": None,
+        },
+        "u_cell_size": {
+            "enabled": True,
+            "group": "2- Core cell size",
+            "label": "Easting (m)",
+            "main": True,
+            "value": 25.0,
+        },
+        "v_cell_size": {
+            "enabled": True,
+            "group": "2- Core cell size",
+            "label": "Northing (m)",
+            "main": True,
+            "value": 25.0,
+        },
+        "w_cell_size": {
+            "enabled": True,
+            "group": "2- Core cell size",
+            "label": "Vertical (m)",
+            "main": True,
+            "value": 25.0,
+        },
+        "horizontal_padding": {
+            "enabled": True,
+            "group": "3- Padding distance",
+            "label": "Horizontal (m)",
+            "main": True,
+            "value": 1000.0,
+        },
+        "vertical_padding": {
+            "enabled": True,
+            "group": "3- Padding distance",
+            "label": "Vertical (m)",
+            "main": True,
+            "value": 1000.0,
+        },
+        "depth_core": {
+            "enabled": True,
+            "group": "1- Core",
+            "label": "Minimum Depth (m)",
+            "main": True,
+            "value": 500.0,
+        },
+        "ga_group_name": {
+            "enabled": True,
+            "group": None,
+            "label": "Name:",
+            "value": "Octree_Mesh",
+        },
+        "run_command": "geoapps.octree_creation.driver",
+    }
+)
 
 template_dict = {
     "object": {
@@ -137,6 +131,8 @@ template_dict = {
         "value": 1000.0,
     },
 }
+
+validations = {}
 
 app_initializer = {
     "geoh5": "../../assets/FlinFlon.geoh5",
