+++ geoapps/peak_finder/constants.py
@@ -39,6 +39,12 @@ defaults = {
     "group_auto": True,
     "center": None,
     "width": None,
+    "run_command": "geoapps.peak_finder.driver",
+    "run_command_boolean": False,
+    "monitoring_directory": None,
+    "workspace_geoh5": None,
+    "conda_environment": "geoapps",
+    "conda_environment_boolean": False,
 }
 
 default_ui_json = deepcopy(base_ui_json)
@@ -148,6 +154,7 @@ default_ui_json.update(
         },
         "center": None,
         "width": None,
+        "run_command": "geoapps.octree_creation.driver",
     }
 )
