+++ geoapps/utils/utils.py
@@ -1137,6 +1137,374 @@ class LineDataDerivatives:
         return deriv
 
 
+def default_groups_from_property_group(property_group, start_index=0):
+
+    _default_channel_groups = {
+        "early": {"label": ["early"], "color": "#0000FF", "channels": []},
+        "middle": {"label": ["middle"], "color": "#FFFF00", "channels": []},
+        "late": {"label": ["late"], "color": "#FF0000", "channels": []},
+        "early + middle": {
+            "label": ["early", "middle"],
+            "color": "#00FFFF",
+            "channels": [],
+        },
+        "early + middle + late": {
+            "label": ["early", "middle", "late"],
+            "color": "#008000",
+            "channels": [],
+        },
+        "middle + late": {
+            "label": ["middle", "late"],
+            "color": "#FFA500",
+            "channels": [],
+        },
+    }
+
+    parent = property_group.parent
+
+    data_list = [
+        parent.workspace.get_entity(uid)[0] for uid in property_group.properties
+    ]
+
+    start = start_index
+    end = len(data_list)
+    block = int((end - start) / 3)
+    ranges = {
+        "early": np.arange(start, start + block).tolist(),
+        "middle": np.arange(start + block, start + 2 * block).tolist(),
+        "late": np.arange(start + 2 * block, end).tolist(),
+    }
+
+    channel_groups = {}
+    for ii, (key, default) in enumerate(_default_channel_groups.items()):
+        prop_group = parent.find_or_create_property_group(name=key)
+        prop_group.properties = []
+
+        for val in default["label"]:
+            for ind in ranges[val]:
+                prop_group.properties += [data_list[ind].uid]
+
+        channel_groups[prop_group.name] = {
+            "data": prop_group.uid,
+            "color": default["color"],
+            "label": [ii + 1],
+            "properties": prop_group.properties,
+        }
+
+    return channel_groups
+
+
+def find_anomalies(
+    locations,
+    line_indices,
+    channels,
+    channel_groups,
+    smoothing=1,
+    use_residual=False,
+    data_normalization=[1],
+    min_amplitude=25,
+    min_value=-np.inf,
+    min_width=200,
+    max_migration=50,
+    min_channels=3,
+    minimal_output=False,
+    return_profile=False,
+):
+    """
+    Find all anomalies along a line profile of data.
+    Anomalies are detected based on the lows, inflection points and a peaks.
+    Neighbouring anomalies are then grouped and assigned a channel_group label.
+
+    :param: :obj:`geoh5py.objects.Curve`
+        Curve object containing data.
+    :param: list
+        List of Data channels
+    :param: array of int or bool
+        Array defining a line of data from the input Curve object
+
+
+    :return: list of dict
+    """
+    profile = LineDataDerivatives(
+        locations=locations[line_indices], smoothing=smoothing, residual=use_residual
+    )
+    locs = profile.locations_resampled
+    if data_normalization == "ppm":
+        data_normalization = [1e-6]
+
+    if locs is None:
+        return {}
+
+    xy = np.c_[profile.interp_x(locs), profile.interp_y(locs)]
+    angles = np.arctan2(xy[1:, 1] - xy[:-1, 1], xy[1:, 0] - xy[:-1, 0])
+    angles = np.r_[angles[0], angles].tolist()
+    azimuth = (450.0 - np.rad2deg(running_mean(angles, width=5))) % 360.0
+    anomalies = {
+        "channel": [],
+        "start": [],
+        "inflx_up": [],
+        "peak": [],
+        "peak_values": [],
+        "inflx_dwn": [],
+        "end": [],
+        "amplitude": [],
+        "group": [],
+        "channel_group": [],
+    }
+    data_uid = list(channels.keys())
+    property_groups = [pg for pg in channel_groups.values()]
+    group_prop_size = np.r_[[len(grp["properties"]) for grp in channel_groups.values()]]
+    for cc, (uid, params) in enumerate(channels.items()):
+        if "values" not in list(params.keys()):
+            continue
+
+        values = params["values"][line_indices].copy()
+        profile.values = values
+        values = profile.values_resampled
+        dx = profile.derivative(order=1)
+        ddx = profile.derivative(order=2)
+        peaks = np.where(
+            (np.diff(np.sign(dx)) != 0) & (ddx[1:] < 0) & (values[:-1] > min_value)
+        )[0]
+        lows = np.where(
+            (np.diff(np.sign(dx)) != 0) & (ddx[1:] > 0) & (values[:-1] > min_value)
+        )[0]
+        lows = np.r_[0, lows, locs.shape[0] - 1]
+        up_inflx = np.where(
+            (np.diff(np.sign(ddx)) != 0) & (dx[1:] > 0) & (values[:-1] > min_value)
+        )[0]
+        dwn_inflx = np.where(
+            (np.diff(np.sign(ddx)) != 0) & (dx[1:] < 0) & (values[:-1] > min_value)
+        )[0]
+
+        if len(peaks) == 0 or len(lows) < 2 or len(up_inflx) < 2 or len(dwn_inflx) < 2:
+            continue
+
+        for peak in peaks:
+            ind = np.median(
+                [0, lows.shape[0] - 1, np.searchsorted(locs[lows], locs[peak]) - 1]
+            ).astype(int)
+            start = lows[ind]
+            ind = np.median(
+                [0, lows.shape[0] - 1, np.searchsorted(locs[lows], locs[peak])]
+            ).astype(int)
+            end = np.min([locs.shape[0] - 1, lows[ind]])
+            ind = np.median(
+                [
+                    0,
+                    up_inflx.shape[0] - 1,
+                    np.searchsorted(locs[up_inflx], locs[peak]) - 1,
+                ]
+            ).astype(int)
+            inflx_up = up_inflx[ind]
+            ind = np.median(
+                [
+                    0,
+                    dwn_inflx.shape[0] - 1,
+                    np.searchsorted(locs[dwn_inflx], locs[peak]),
+                ]
+            ).astype(int)
+            inflx_dwn = np.min([locs.shape[0] - 1, dwn_inflx[ind] + 1])
+            # Check amplitude and width thresholds
+            delta_amp = (
+                np.abs(
+                    np.min([values[peak] - values[start], values[peak] - values[end]])
+                )
+                / (np.std(values) + 2e-32)
+            ) * 100.0
+            delta_x = locs[end] - locs[start]
+            amplitude = np.sum(np.abs(values[start:end])) * profile.sampling
+            if (delta_amp > min_amplitude) & (delta_x > min_width):
+                anomalies["channel"] += [cc]
+                anomalies["start"] += [start]
+                anomalies["inflx_up"] += [inflx_up]
+                anomalies["peak"] += [peak]
+                anomalies["peak_values"] += [values[peak]]
+                anomalies["inflx_dwn"] += [inflx_dwn]
+                anomalies["amplitude"] += [amplitude]
+                anomalies["end"] += [end]
+                anomalies["group"] += [-1]
+                anomalies["channel_group"] += [
+                    [
+                        key
+                        for key, channel_group in enumerate(channel_groups.values())
+                        if uid in channel_group["properties"]
+                    ]
+                ]
+
+    if len(anomalies["peak"]) == 0:
+        if return_profile:
+            return {}, profile
+        else:
+            return {}
+
+    groups = []
+
+    # Re-cast as numpy arrays
+    for key, values in anomalies.items():
+        if key == "channel_group":
+            continue
+        anomalies[key] = np.hstack(values)
+
+    group_id = -1
+    peaks_position = locs[anomalies["peak"]]
+    for ii in range(peaks_position.shape[0]):
+        # Skip if already labeled
+        if anomalies["group"][ii] != -1:
+            continue
+
+        group_id += 1  # Increment group id
+        dist = np.abs(peaks_position[ii] - peaks_position)
+        # Find anomalies across channels within horizontal range
+        near = np.where((dist < max_migration) & (anomalies["group"] == -1))[0]
+        # Reject from group if channel gap > 1
+        u_gates, u_count = np.unique(anomalies["channel"][near], return_counts=True)
+        if len(u_gates) > 1 and np.any((u_gates[1:] - u_gates[:-1]) > 2):
+            cutoff = u_gates[np.where((u_gates[1:] - u_gates[:-1]) > 2)[0][0]]
+            near = near[anomalies["channel"][near] <= cutoff]  # Remove after cutoff
+        # Check for multiple nearest peaks on single channel
+        # and keep the nearest
+        u_gates, u_count = np.unique(anomalies["channel"][near], return_counts=True)
+        for gate in u_gates[np.where(u_count > 1)]:
+            mask = np.ones_like(near, dtype="bool")
+            sub_ind = anomalies["channel"][near] == gate
+            sub_ind[np.where(sub_ind)[0][np.argmin(dist[near][sub_ind])]] = False
+            mask[sub_ind] = False
+            near = near[mask]
+
+        score = np.zeros(len(channel_groups))
+        for ids in near:
+            score[anomalies["channel_group"][ids]] += 1
+
+        # Find groups with largest channel overlap
+        max_scores = np.where(score == score.max())[0]
+        # Keep the group with less properties
+        in_group = max_scores[
+            np.argmax(score[max_scores] / group_prop_size[max_scores])
+        ]
+        if score[in_group] < min_channels:
+            continue
+
+        channel_group = property_groups[in_group]
+        # Remove anomalies not in group
+        mask = [
+            data_uid[anomalies["channel"][id]] in channel_group["properties"]
+            for id in near
+        ]
+        near = near[mask, ...]
+        if len(near) == 0:
+            continue
+        anomalies["group"][near] = group_id
+        gates = anomalies["channel"][near]
+        cox = anomalies["peak"][near]
+        inflx_dwn = anomalies["inflx_dwn"][near]
+        inflx_up = anomalies["inflx_up"][near]
+        cox_sort = np.argsort(locs[cox])
+        azimuth_near = azimuth[cox]
+        dip_direction = azimuth[cox[0]]
+
+        if (
+            anomalies["peak_values"][near][cox_sort][0]
+            < anomalies["peak_values"][near][cox_sort][-1]
+        ):
+            dip_direction = (dip_direction + 180) % 360.0
+
+        migration = np.abs(locs[cox[cox_sort[-1]]] - locs[cox[cox_sort[0]]])
+        skew = (locs[cox][cox_sort[0]] - locs[inflx_up][cox_sort]) / (
+            locs[inflx_dwn][cox_sort] - locs[cox][cox_sort[0]] + 1e-8
+        )
+        skew[azimuth_near[cox_sort] > 180] = 1.0 / (
+            skew[azimuth_near[cox_sort] > 180] + 1e-2
+        )
+        # Change skew factor from [-100, 1]
+        flip_skew = skew < 1
+        skew[flip_skew] = 1.0 / (skew[flip_skew] + 1e-2)
+        skew = 1.0 - skew
+        skew[flip_skew] *= -1
+        values = anomalies["peak_values"][near] * np.prod(data_normalization)
+        amplitude = np.sum(anomalies["amplitude"][near])
+        times = [
+            channel["time"]
+            for ii, channel in enumerate(channels.values())
+            if (ii in list(gates) and "time" in channel.keys())
+        ]
+        linear_fit = None
+
+        if len(times) > 2 and len(cox) > 0:
+            times = np.hstack(times)[values > 0]
+            if len(times) > 2:
+                # Compute linear trend
+                A = np.c_[np.ones_like(times), times]
+                y0, slope = np.linalg.solve(
+                    np.dot(A.T, A), np.dot(A.T, np.log(values[values > 0]))
+                )
+                linear_fit = [y0, slope]
+
+        group = {
+            "channels": gates,
+            "start": anomalies["start"][near],
+            "inflx_up": anomalies["inflx_up"][near],
+            "peak": cox,
+            "cox": np.mean(
+                np.c_[
+                    profile.interp_x(locs[cox[cox_sort[0]]]),
+                    profile.interp_y(locs[cox[cox_sort[0]]]),
+                    profile.interp_z(locs[cox[cox_sort[0]]]),
+                ],
+                axis=0,
+            ),
+            "inflx_dwn": anomalies["inflx_dwn"][near],
+            "end": anomalies["end"][near],
+            "azimuth": dip_direction,
+            "migration": migration,
+            "amplitude": amplitude,
+            "channel_group": channel_group,
+            "linear_fit": linear_fit,
+        }
+        if minimal_output:
+
+            group["skew"] = np.mean(skew)
+            group["inflx_dwn"] = np.c_[
+                profile.interp_x(locs[inflx_dwn]),
+                profile.interp_y(locs[inflx_dwn]),
+                profile.interp_z(locs[inflx_dwn]),
+            ]
+            group["inflx_up"] = np.c_[
+                profile.interp_x(locs[inflx_up]),
+                profile.interp_y(locs[inflx_up]),
+                profile.interp_z(locs[inflx_up]),
+            ]
+            start = anomalies["start"][near]
+            group["start"] = np.c_[
+                profile.interp_x(locs[start]),
+                profile.interp_y(locs[start]),
+                profile.interp_z(locs[start]),
+            ]
+
+            end = anomalies["end"][near]
+            group["peaks"] = np.c_[
+                profile.interp_x(locs[cox]),
+                profile.interp_y(locs[cox]),
+                profile.interp_z(locs[cox]),
+            ]
+
+            group["end"] = np.c_[
+                profile.interp_x(locs[end]),
+                profile.interp_y(locs[end]),
+                profile.interp_z(locs[end]),
+            ]
+
+        else:
+            group["peak_values"] = values
+
+        groups += [group]
+
+    if return_profile:
+        return groups, profile
+    else:
+        return groups
+
+
 def tensor_2_block_model(workspace, mesh, name=None, parent=None, data={}):
     """
     Function to convert a tensor mesh from :obj:`~discretize.TensorMesh` to
